[refactor] refactor package name
1. refactor package name (diarylike to diary_like)
