+++ back/demo/src/main/java/OOTD/demo/diary_like/dto/PostDiaryLikeReqDTO.java
@@ -1,4 +1,4 @@
-package OOTD.demo.diarylike.dto;
+package OOTD.demo.diary_like.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
